// src/interfaces/keycloak-user.interface.ts
export interface IKeycloakUser {
  keycloakId: string;
  email_verified: boolean;
  name: string;
  preferred_username: string;
  given_name: string;
  family_name: string;
  email: string;
  realm_access: {
    roles: string[];
  };
  resource_access: {
    [key: string]: { roles: string[] };
  };
  customerIds: number[];
  clientCodes: string[];
  roles: string[];
}